move alphabet component to /game page and put alphabet generation into separate file
