+++ src/pages/Result.jsx
@@ -3,7 +3,6 @@ import { Link } from "react-router";
 export default function Result() {
   return (
     <>
-      <h1>Result</h1>
       <Link to="/game">Start new game</Link>
       <Link to="/">Back home</Link>
     </>
